Readded MultiArchTest.groovy and updated testutils to pass script downto test objects
