+++ vars/TestUtils.groovy
@@ -4,22 +4,24 @@ import com.redhat.multiarch.ci.task.*
 
 class TestUtils {
   static ProvisioningConfig config = null
-  
-  static ProvisioningConfig getProvisioningConfig(script) {
+
+  static ProvisioningConfig getProvisioningConfig(Workflow script) {
     if (!config) config = new ProvisioningConfig(script.params, script.env)
     config
   }
 
   /**
-   * Runs @test on the multi-arch capable provisioner container,
-   * an runs @onTestFailure if it encounter an Exception.
+   * Runs @test on a multi-arch provisioned host for the specified arch.
+   * Runs @onTestFailure if it encounters an Exception.
    *
+   * @param script WorkflowScript that the test will run in.
    * @param arch String specifying the arch to run tests on.
    * @param config ProvisioningConfig Configuration for provisioning.
    * @param test Closure that takes the Slave used by the test.
    * @param onTestFailure Closure that take the Slave used by the test and the Exception that occured.
    */
   static def runTest(
+    WorkflowScript script,
     String arch,
     ProvisioningConfig config,
     Closure test,
@@ -29,8 +31,10 @@ class TestUtils {
 
 
   /**
-   * Run closure body on a multi-arch provisioned host for each arch in arches param.
+   * Runs @test on a multi-arch provisioned host for each arch in arches param.
+   * Runs @onTestFailure if it encounters an Exception.
    *
+   * @param script WorkflowScript that this test will run in.
    * @param arches List<String> specifying the arches to run single host tests on.
    * @param config ProvisioningConfig Configuration for provisioning.
    * @param test Closure that takes the Slave used by the test.
@@ -42,21 +46,6 @@ class TestUtils {
     ProvisioningConfig config,
     Closure test,
     Closure onTestFailure) {
-    // Create arch Tasks to parallelize test
-    def parallelTasks = []
-    for (arch in arches) {
-      parallelTasks.push(new Task(name: arch, params: [ arch: arch ]))
-    }
-
-    // Run single host test in parallel on each arch
-    script.parallel Task.parallelizeTaskList(
-      parallelTasks,
-      { params ->
-        def arch = params.arch
-        return {
-          (new Test(arch, config, test, onTestFailure)).run()
-        }
-      }
-    )
+    (new MultiArchTest(script, arches, config, test, onTestFailure)).run()
   }
 }
